requirements.txt
astroid==2.1.0
autopep8==1.4.3
coverage==4.5.2
Django==2.1.3
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
<<<<<<< HEAD
flake8==3.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
=======
<<<<<<< HEAD
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
=======
>>>>>>> 018e53839fcbac4df69aeb430eb23796816b1fb1
>>>>>>> chore(article): test for article
PyJWT==1.7.0
pylint==2.2.2
pytz==2018.7
six==1.11.0
wrapt==1.10.11
